requirements.txt
aiohttp==0.21.2
aiohttp-debugtoolbar==0.1.1
aiohttp-jinja2==0.7.0
aiopg==0.9.2
aioredis==0.2.5
chardet==2.3.0
Django==1.9.4
django-debug-toolbar==1.4
django-filter==0.12.0
django-redis-cache==1.6.5
djangorestframework==3.3.2
hiredis==0.2.0
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
Pillow==3.1.1
psycopg2==2.6.1
redis==2.10.5
sqlparse==0.1.19
